Release 6.6.2

+++ module_common/src/main/java/com/sensorsdata/analytics/android/sdk/data/SensorsDataDBHelper.java
@@ -17,6 +17,7 @@
 package com.sensorsdata.analytics.android.sdk.data;
 
 import android.content.Context;
+import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
@@ -26,10 +27,10 @@ import com.sensorsdata.analytics.android.sdk.data.adapter.DbParams;
 class SensorsDataDBHelper extends SQLiteOpenHelper {
     private static final String TAG = "SA.SQLiteOpenHelper";
     private static final String CREATE_EVENTS_TABLE =
-            String.format("CREATE TABLE %s (_id INTEGER PRIMARY KEY AUTOINCREMENT, %s TEXT NOT NULL, %s INTEGER NOT NULL);", DbParams.TABLE_EVENTS, DbParams.KEY_DATA, DbParams.KEY_CREATED_AT);
+            String.format("CREATE TABLE IF NOT EXISTS %s (_id INTEGER PRIMARY KEY AUTOINCREMENT, %s TEXT NOT NULL, %s INTEGER NOT NULL, %s INTEGER NOT NULL DEFAULT 0);", DbParams.TABLE_EVENTS, DbParams.KEY_DATA, DbParams.KEY_CREATED_AT, DbParams.KEY_IS_INSTANT_EVENT);
     private static final String EVENTS_TIME_INDEX =
             String.format("CREATE INDEX IF NOT EXISTS time_idx ON %s (%s);", DbParams.TABLE_EVENTS, DbParams.KEY_CREATED_AT);
-    private static final String CHANNEL_EVENT_PERSISTENT_TABLE = String.format("CREATE TABLE %s (%s TEXT PRIMARY KEY, %s INTEGER)",
+    private static final String CHANNEL_EVENT_PERSISTENT_TABLE = String.format("CREATE TABLE IF NOT EXISTS %s (%s TEXT PRIMARY KEY, %s INTEGER)",
             DbParams.TABLE_CHANNEL_PERSISTENT, DbParams.KEY_CHANNEL_EVENT_NAME, DbParams.KEY_CHANNEL_RESULT);
 
     SensorsDataDBHelper(Context context) {
@@ -39,17 +40,30 @@ class SensorsDataDBHelper extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase db) {
         SALog.i(TAG, "Creating a new Sensors Analytics DB");
-
-        db.execSQL(CREATE_EVENTS_TABLE);
-        db.execSQL(EVENTS_TIME_INDEX);
-        db.execSQL(CHANNEL_EVENT_PERSISTENT_TABLE);
+        createTable(db);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        SALog.i(TAG, "Upgrading app, replacing Sensors Analytics DB");
+        SALog.i(TAG, "Upgrading app, replacing Sensors Analytics DB, oldVersion:" + oldVersion + ", newVersion:" + newVersion);
+        try {
+            if (oldVersion < 4) {
+                db.execSQL(String.format("DROP TABLE IF EXISTS %s", DbParams.TABLE_EVENTS));
+            }
+            createTable(db);
+            if (oldVersion >= 4 && oldVersion <= 5) {
+                //不存在则创建字段，避免数据库升级，又降级，又升级有此字段导致的字段重复添加异常
+                if (!checkColumnExist(db, DbParams.TABLE_EVENTS, DbParams.KEY_IS_INSTANT_EVENT)) {
+                    String sql = "ALTER TABLE " + DbParams.TABLE_EVENTS + " ADD COLUMN  " + DbParams.KEY_IS_INSTANT_EVENT + " INTEGER NOT NULL DEFAULT 0";
+                    db.execSQL(sql);
+                }
+            }
+        } catch (Exception e) {
+            SALog.printStackTrace(e);
+        }
+    }
 
-        db.execSQL(String.format("DROP TABLE IF EXISTS %s", DbParams.TABLE_EVENTS));
+    private void createTable(SQLiteDatabase db) {
         db.execSQL(CREATE_EVENTS_TABLE);
         db.execSQL(EVENTS_TIME_INDEX);
         db.execSQL(CHANNEL_EVENT_PERSISTENT_TABLE);
@@ -57,5 +71,36 @@ class SensorsDataDBHelper extends SQLiteOpenHelper {
 
     @Override
     public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+    }
+
+    /**
+     * 检查某表中某列是否存在
+     *
+     * @param db 数据库
+     * @param tableName 表名
+     * @param columnName 列名
+     * @return 列表某列是否存在
+     */
+    private boolean checkColumnExist(SQLiteDatabase db, String tableName
+            , String columnName) {
+        boolean result = false;
+        Cursor cursor = null;
+        try {
+            //查询一行
+            cursor = db.rawQuery("SELECT * FROM " + tableName + " LIMIT 0"
+                    , null);
+            result = cursor != null && cursor.getColumnIndex(columnName) != -1;
+        } catch (Exception e) {
+            SALog.printStackTrace(e);
+        } finally {
+            try {
+                if (null != cursor && !cursor.isClosed()) {
+                    cursor.close();
+                }
+            } catch (Exception e) {
+                SALog.printStackTrace(e);
+            }
+        }
+        return result;
     }
 }
